Update README.md
## World_Layoffs

### Project Overview
This data analysis project aims at providing insight on world layoff between march 2020 and march 2023.  By analyzing this data, we aim to identify trends and gain deeper understanding of the world layoff within the said time.

### Data source
The primary dataset used for this analysis is “world layoffs” dataset. It contains detailed information of layoffs that happened in various US companies from March 2020 to March 2023.

### Tools
The tools used for this analysis are 
-SQL for data cleaning and data analysis
- Tableau for reporting

### Cleaning and Preparation
The data preparation stage covered;
1.	Data loading and inspection
2.	Removing duplicate values
3.	Data standardization
4.	Null values or blank values

### Exploratory Data Analysis
EDA involved working with the dataset to answer the following questions
1.	What is the max and min laid off